standard step corrected
The standard step, or size of each file piece, will be fixed back to 4096.
The reasoning is that the program was originally designed to split text files for the iPod. This size of standard split is the maximum note size that the iPod will handle. The old step makes 4MB pieces. Because of the extremely large file size of a test, 4KB pieces would have been too much for the test system. In other words it was a temporary change.